[mkt test] move wait for broadcast fns to be near create config
